Modified Contact us and Reset Password api's to use different routes
